+++ gamemodes/terrortown/gamemode/client/cl_player_ext.lua
@@ -199,6 +199,16 @@ function plymeta:GetVoiceColor()
     return self.voiceColor or INNOCENT.color
 end
 
+-- Creates an elliptic nick for a given length.
+-- @param number width The maximum width that should be used to limit the nick
+-- @param[default="DefaultBold"] string font The font ID
+-- @param[default=1.0] number scale The UI scale factor
+-- @return string The length limited nick
+-- @realm client
+function plymeta:NickElliptic(width, font, scale)
+    return draw.GetLimitedLengthText(self:Nick(), width, font, "...", scale)
+end
+
 ---
 -- @hook
 -- @param Player ply The player to update the animation info for.